refactor: stages::Pipeline --> reth_pipeline::pipeline::Pipeline
makes the separation between the pipeline and the stages cleaner

also lets us cleanly introduce a pipeline builder and relevant configs
