+++ crates/reth/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "reth-pipeline"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/paradigmxyz/reth"
+readme = "README.md"
+
+[dependencies]
+# reth libs
+reth-stages = { path = "../stages" }
+reth-primitives = { path = "../primitives" }
+reth-interfaces = { path = "../interfaces" }
+reth-executor = { path = "../executor" }
+reth-rlp = { path = "../common/rlp" }
+reth-db = { path = "../storage/db" }
+reth-provider = { path = "../storage/provider" }
+reth-metrics-derive = { path = "../metrics/metrics-derive" }
+
+# async
+tokio = { version = "1.21.2", features = ["sync"] }
+
+async-trait = "0.1.57"
+thiserror = "1.0.37"
+tracing = "0.1.36"
+aquamarine = "0.1.12"
+metrics = "0.20.1"
+futures-util = "0.3.25"
+itertools = "0.10.5"
+rayon = "1.6.0"
+
+[dev-dependencies]
+# reth
+reth-db = { path = "../storage/db", features = ["test-utils", "mdbx"] }
+reth-interfaces = { path = "../interfaces", features = ["test-utils"] }
+reth-downloaders = { path = "../net/downloaders" }
+reth-eth-wire = { path = "../net/eth-wire" }                            # TODO(onbjerg): We only need this for [BlockBody]
+tokio = { version = "*", features = ["rt", "sync", "macros"] }
+tokio-stream = "0.1.10"
+tempfile = "3.3.0"
+assert_matches = "1.5.0"
+rand = "0.8.5"
