+++ crates/stages/src/util.rs
@@ -1,34 +1,35 @@
-pub(crate) mod opt {
+/// Helpers for working with Options
+pub mod opt {
     use tokio::sync::mpsc::{error::SendError, Sender};
 
     /// Get an [Option] with the maximum value, compared between the passed in value and the inner
     /// value of the [Option]. If the [Option] is `None`, then an option containing the passed in
     /// value will be returned.
-    pub(crate) fn max<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
+    pub fn max<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
         a.map_or(Some(b), |v| Some(std::cmp::max(v, b)))
     }
 
     /// Get an [Option] with the minimum value, compared between the passed in value and the inner
     /// value of the [Option]. If the [Option] is `None`, then an option containing the passed in
     /// value will be returned.
-    pub(crate) fn min<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
+    pub fn min<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
         a.map_or(Some(b), |v| Some(std::cmp::min(v, b)))
     }
 
     /// The producing side of a [tokio::mpsc] channel that may or may not be set.
-    #[derive(Default, Clone)]
-    pub(crate) struct MaybeSender<T> {
+    #[derive(Debug, Default, Clone)]
+    pub struct MaybeSender<T> {
         inner: Option<Sender<T>>,
     }
 
     impl<T> MaybeSender<T> {
         /// Create a new [MaybeSender]
-        pub(crate) fn new(sender: Option<Sender<T>>) -> Self {
+        pub fn new(sender: Option<Sender<T>>) -> Self {
             Self { inner: sender }
         }
 
         /// Send a value over the channel if an internal sender has been set.
-        pub(crate) async fn send(&self, value: T) -> Result<(), SendError<T>> {
+        pub async fn send(&self, value: T) -> Result<(), SendError<T>> {
             if let Some(rx) = &self.inner {
                 rx.send(value).await
             } else {
@@ -37,7 +38,7 @@ pub(crate) mod opt {
         }
 
         /// Set or unset the internal sender.
-        pub(crate) fn set(&mut self, sender: Option<Sender<T>>) {
+        pub fn set(&mut self, sender: Option<Sender<T>>) {
             self.inner = sender;
         }
     }
